Compute intersections based on connectivity
In a circuit, a junction should only be added if two intersecting lines
are connected.  This work determines that by giving every junction a
"voltage".  All junctions with the same voltage are a part of the same
subgraph.  When a junction is added, it is given a unique voltage.  When
two junctions are connected, they are set to the lesser of their
voltages.  This change is propagated to both of their entire subgraphs.

This also changes the junction maps to be Rc<RefCell<_>>.  This is
because I didn't realize that RefCell<_> isn't heap allocated, which
caused there to be desyncs in a Line's perception of the junction map
when that map was mutated (e.g. when the voltage was changed).

gdext was also upgraded in an attempt to fix a hot reloading issue,
although it didn't actually.